+++ python/lsst/ts/mtdome/mock_llc/mock_motion/shutter_motion.py
@@ -1,6 +1,6 @@
 # This file is part of ts_mtdome.
 #
-# Developed for the Vera Rubin Observatory Telescope and Site Systems.
+# Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
@@ -30,7 +30,7 @@ __all__ = [
 import logging
 import math
 
-from lsst.ts.idl.enums.MTDome import MotionState
+from lsst.ts.xml.enums.MTDome import MotionState
 
 from ...enums import InternalMotionState
 from .base_llc_motion_without_crawl import BaseLlcMotionWithoutCrawl