+++ python/lsst/ts/mtdome/mock_llc/moncs.py
@@ -1,6 +1,6 @@
 # This file is part of ts_mtdome.
 #
-# Developed for the Vera Rubin Observatory Telescope and Site Systems.
+# Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
@@ -24,7 +24,7 @@ __all__ = ["MoncsStatus", "NUM_MON_SENSORS"]
 import logging
 
 import numpy as np
-from lsst.ts.idl.enums.MTDome import MotionState
+from lsst.ts.xml.enums.MTDome import MotionState
 
 from ..enums import InternalMotionState
 from .base_mock_llc import BaseMockStatus