+++ tests/test_lcs.py
@@ -1,6 +1,6 @@
 # This file is part of ts_mtdome.
 #
-# Developed for the Vera Rubin Observatory Telescope and Site Systems.
+# Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
@@ -23,13 +23,13 @@ import unittest
 
 import numpy as np
 from lsst.ts import mtdome
-from lsst.ts.idl.enums.MTDome import MotionState
 from lsst.ts.mtdome.mock_llc.lcs import (
     CURRENT_PER_MOTOR,
     NUM_LOUVERS,
     NUM_MOTORS_PER_LOUVER,
-    TOTAL_POWER,
 )
+from lsst.ts.mtdome.power_management.power_draw_constants import LOUVERS_POWER_DRAW
+from lsst.ts.xml.enums.MTDome import MotionState
 
 START_TAI = 10001.0
 
@@ -86,7 +86,7 @@ class LcsTestCase(unittest.IsolatedAsyncioTestCase):
                 if lcs_status["status"]["status"][index] == MotionState.MOVING:
                     assert driveCurrentActualMotor1 == CURRENT_PER_MOTOR
                     assert driveCurrentActualMotor2 == CURRENT_PER_MOTOR
-                    assert lcs_status["powerDraw"] == TOTAL_POWER
+                    assert lcs_status["powerDraw"] == LOUVERS_POWER_DRAW
             else:
                 assert driveCurrentActualMotor1 == 0.0
                 assert driveCurrentActualMotor2 == 0.0