+++ tests/test_lwscs.py
@@ -1,6 +1,6 @@
 # This file is part of ts_mtdome.
 #
-# Developed for the Vera Rubin Observatory Telescope and Site Systems.
+# Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
@@ -25,8 +25,9 @@ import unittest
 import pytest
 from expected_state import ExpectedState
 from lsst.ts import mtdome
-from lsst.ts.idl.enums.MTDome import MotionState
-from lsst.ts.mtdome.mock_llc.lwscs import CURRENT_PER_MOTOR, NUM_MOTORS, TOTAL_POWER
+from lsst.ts.mtdome.mock_llc.lwscs import CURRENT_PER_MOTOR, NUM_MOTORS
+from lsst.ts.mtdome.power_management.power_draw_constants import LWS_POWER_DRAW
+from lsst.ts.xml.enums.MTDome import MotionState
 
 START_TAI = 10001.0
 MIN_POSITION = 0
@@ -103,7 +104,7 @@ class LwscsTestCase(unittest.IsolatedAsyncioTestCase):
             MotionState.MOVING,
         ]:
             expected_drive_current = [CURRENT_PER_MOTOR] * NUM_MOTORS
-            expected_power_draw = TOTAL_POWER
+            expected_power_draw = LWS_POWER_DRAW
         assert expected_drive_current == self.lwscs.llc_status["driveCurrentActual"]
         assert expected_power_draw == self.lwscs.llc_status["powerDraw"]
 
@@ -339,7 +340,7 @@ class LwscsTestCase(unittest.IsolatedAsyncioTestCase):
         await self.verify_halt(
             start_tai=2.0,
             expected_states=expected_states,
-            command="stopEl",
+            command=mtdome.CommandName.STOP_EL,
         )
 
     async def test_stationary(self) -> None: